sentinel/pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "Sentinel - Advanced system metrics collector with time-series storage and pattern detection"
authors = [
    {name = "coff33ninja", email = "coff33ninja69@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
keywords = ["system-monitoring", "metrics", "performance", "telemetry", "time-series"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "psutil>=5.9.8",
    "GPUtil>=1.4.0",
    "py-cpuinfo>=9.0.0",
    "wmi>=1.5.1; sys_platform == 'win32'",
    "aiosqlite>=0.19.0",
    "aiofiles>=23.2.1",
    "pydantic>=2.5.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "google-generativeai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
sysmetrics = "cli.main:cli"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["collectors*", "storage*", "aggregator*", "patterns*", "cli*", "utils*"]
exclude = ["tests*", "data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "A", "C4", "DTZ", "T10", "EM", "ISC", "ICN", "G", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "ARG", "PLR2004"]
"__init__.py" = ["F401"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", ".venv/*", "setup.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
